Jest integration/server testing complete for now.

+++ db-cassandra/index.js
@@ -1,19 +1,6 @@
 const cassandra = require('cassandra-driver');
 const client = require('../config/cassConnection');
 
-//for testing connection
-// client.connect()
-//   .then(() => {
-//     console.log('Connected to cluster with %d host(s): %j', client.hosts.length, client.hosts.keys());
-//     console.log('Keyspaces: %j', Object.keys(client.metadata.keyspaces));
-//     console.log('Shutting down');
-//     return client.shutdown();
-//   })
-//   .catch((err) => {
-//     console.error('There was an error when connecting', err);
-//     return client.shutdown();
-//   });
-
 const validCols = {
   rest_id: true,
   address: true,
@@ -49,12 +36,24 @@ const validCols = {
 };
 
 const validColKeys = Object.keys(validCols);
+const reg = new RegExp('^[0-9]+$');
 
 const checkParams = (req) => {
-  const rest_id = parseInt(req.params.restaurantId, 10);
-  if (rest_id !== rest_id) {
-    throw new Error('bad request: 400');
-    return 'error: invalid rest_id';
+  const paramsID = req.params.restaurantId
+  const rest_id = parseInt(paramsID, 10);
+  if (rest_id !== rest_id || !reg.test(paramsID)) {
+    return 'bad request: 400';
+  } else {
+    return rest_id;
+  }
+};
+
+const checkParamSelect = (req) => {
+  const paramsID = req.params.restaurantId
+  const rest_id = parseInt(paramsID, 10);
+  if (rest_id !== rest_id || !reg.test(paramsID)) {
+    throw new Error('404');
+    return [];
   } else {
     return rest_id;
   }
@@ -96,10 +95,6 @@ const selectByID = async (req, res, next) => {
     const resultSet = await client.execute(querySelectRestID, [rest_id], { prepare: true });
     return resultSet;
   } catch(err) {
-    if (err.message === 'bad request: 400') {
-      res.sendStatus(400);
-      return err;
-    }
     return err;
   }
 };
@@ -107,15 +102,15 @@ const selectByID = async (req, res, next) => {
 const updateByID = async (req, res, next) => {
   try {
     const rest_id = checkParams(req);
+    if (rest_id === 'bad request: 400') {
+      throw new Error('bad request: 400');
+    }
     const queryUpdateRestID = `UPDATE overview_by_id SET ${req.body.data.column.colName} = ? WHERE rest_id = ? IF EXISTS`;
     checkUpdateRequestBody(req, res, next);
     const resultSet = await client.execute(queryUpdateRestID, [req.body.data.column.colNew, rest_id], { prepare: true });
     return resultSet;
   } catch(err) {
-    if (err.message === 'bad request: 400') {
-      res.sendStatus(400);
-      return err;
-    } else if (err.message === 'no content: 204') {
+    if (err.message === 'no content: 204') {
       res.sendStatus(204);
       return err;
     }
@@ -132,19 +127,13 @@ const queryInsertRow = 'INSERT INTO overview_by_id (rest_id, address, breakfast,
 const insertIntoDB = async (req, res, next) => {
   try {
     const rest_id = checkParams(req);
-    console.log('Found: insertIntoDB');
-    console.log(req);
+    if (rest_id === 'bad request: 400') {
+      throw new Error('bad request: 400');
+    }
     const postReqData = checkPostRequestBody(req);
     const resultSet = await client.execute(queryInsertRow, postReqData, { prepare: true });
     return resultSet;
   } catch(err) {
-    if (err.message === 'no content: 204') {
-      res.sendStatus(204);
-      return err;
-    } else if (err.message === 'bad request: 400') {
-      res.sendStatus(400);
-      return err;
-    }
     return err;
   }
 };
@@ -152,6 +141,9 @@ const insertIntoDB = async (req, res, next) => {
 const deleteRowDB = async (req, res, next) => {
   try {
     const rest_id = checkParams(req);
+    if (rest_id === 'bad request: 400') {
+      throw new Error('bad request: 400');
+    }
     const queryDeleteRow = `DELETE FROM greyscale_overviews2.overview_by_id WHERE rest_id = ? IF EXISTS`;
     const resultSet = await client.execute(queryDeleteRow, [rest_id], { prepare: true });
     return resultSet;
@@ -160,50 +152,9 @@ const deleteRowDB = async (req, res, next) => {
   }
 };
 
-const testInsertIntoDB = async (client, id, reqBody) => {
-  try {
-    const req = {
-      params: {
-        restaurantId: id
-      }
-    }
-    const rest_id = checkParams(req);
-    const postReqData = checkPostRequestBody(reqBody);
-    const resultSet = await client.execute(queryInsertRow, postReqData, { prepare: true });
-    return resultSet;
-  } catch(err) {
-    if (err.message === 'no content: 204') {
-      res.sendStatus(204);
-      return err;
-    } else if (err.message === 'bad request: 400') {
-      res.sendStatus(400);
-      return err;
-    }
-    return err;
-  }
-}
-
-const testDeleteRowDB = async (client, id) => {
-  try {
-    const req = {
-      params: {
-        restaurantId: id
-      }
-    }
-    const rest_id = checkParams(req);
-    const queryDeleteRow = `DELETE FROM greyscale_overviews2.overview_by_id WHERE rest_id = ? IF EXISTS`;
-    const resultSet = await client.execute(queryDeleteRow, [rest_id], { prepare: true });
-    return resultSet;
-  } catch(err) {
-    return err;
-  }
-}
-
 module.exports = {
   selectByID,
   updateByID,
   insertIntoDB,
   deleteRowDB,
-  testInsertIntoDB,
-  testDeleteRowDB,
 }
